Split of isotp.py
