+++ test/contrib/isotpscan.uts
@@ -6,7 +6,7 @@
 
 = Imports
 import scapy.modules.six as six
-from scapy.contrib.isotp import send_multiple_ext, filter_periodic_packets, scan_extended, scan
+from scapy.contrib.isotp.isotp_scan import send_multiple_ext, filter_periodic_packets, scan_extended, scan
 
 if six.PY3:
     exec(open("test/contrib/automotive/interface_mockup.py").read())
@@ -776,14 +776,14 @@ assert cleanup_interfaces()
 
 = empty tests
 
-from scapy.contrib.isotp import generate_code_output, generate_text_output
+from scapy.contrib.isotp.isotp_scan import generate_code_output, generate_text_output
 
 assert generate_code_output("", None) == ""
 assert generate_text_output("") == "No packets found."
 
 = get_isotp_fc
 
-from scapy.contrib.isotp import get_isotp_fc
+from scapy.contrib.isotp.isotp_scan import get_isotp_fc
 
 # to trigger "noise_ids.append(packet.identifier)"
 a = []
